Split theme api from ops
